+++ cli_tool_audit/__about__.py
@@ -1,0 +1,26 @@
+"""Metadata for cli_tool_audit.
+"""
+
+__all__ = [
+    "__title__",
+    "__name__",
+    "__version__",
+    "__description__",
+    "__author__",
+    "__author_email__",
+    "__keywords__",
+    "__status__",
+    "__license__",
+    "__homepage__",
+]
+
+__title__ = "cli_tool_audit"
+__name__ = "cli_tool_audit"
+__version__ = "1.2.0"
+__description__ = "Audit for existence and version number of cli tools."
+__author__ = "Matthew Martin"
+__author_email__ = "matthewdeanmartin@gmail.com"
+__keywords__ = ["cli tooling", "version numbers"]
+__status__ = "4 - Beta"
+__license__ = "MIT"
+__homepage__ = "https://github.com/matthewdeanmartin/cli_tool_audit"