+++ cli_tool_audit/call_and_compatible.py
@@ -5,24 +5,26 @@ import sys
 import threading
 
 from cli_tool_audit.audit_cache import AuditFacade
+from cli_tool_audit.audit_manager import AuditManager
 from cli_tool_audit.models import CliToolConfig, ToolCheckResult
 
 # Old interface
 
 
 def check_tool_wrapper(
-    tool_info: tuple[str, CliToolConfig, threading.Lock],
+    tool_info: tuple[str, CliToolConfig, threading.Lock, bool],
 ) -> ToolCheckResult:
     """
     Wrapper function for check_tool_availability() that returns a ToolCheckResult object.
 
     Args:
-        tool_info (tuple[str, CliToolConfig]): A tuple containing the tool name and the tool configuration.
+        tool_info (tuple[str, CliToolConfig, threading.Lock, bool]): A tuple containing the tool name, the
+        CliToolConfig object, a lock, and a boolean indicating if the cache is enabled.
 
     Returns:
         ToolCheckResult: A ToolCheckResult object.
     """
-    tool, config, lock = tool_info
+    tool, config, lock, enable_cache = tool_info
 
     if config.if_os and not sys.platform.startswith(config.if_os):
         # This isn't very transparent about what just happened
@@ -39,11 +41,17 @@ def check_tool_wrapper(
             last_modified=None,
         )
 
-    manager = AuditFacade()
     config.name = tool
     config.version_switch = config.version_switch or "--version"
-    with lock:
-        return manager.call_and_check(tool_config=config)
+
+    if enable_cache:
+        cached_manager = AuditFacade()
+        with lock:
+            return cached_manager.call_and_check(tool_config=config)
+
+    manager = AuditManager()
+    return manager.call_and_check(tool_config=config)
+
     # result = cli_availability.check_tool_availability(
     #     tool, config.version_switch or "--version", cast(bool, config.only_check_existence or False)
     # )