+++ cli_tool_audit/models.py
@@ -9,15 +9,20 @@ from typing import Optional
 
 @dataclass
 class ToolCheckResult:
+    """
+    Represents the result of a tool check and version check.
+    """
+
     tool: str
     desired_version: str
     is_needed_for_os: bool
     is_available: bool
     is_snapshot: bool
+    """Snapshot schema"""
     found_version: Optional[str]
     """Same as snapshot_version, the exact text produced by --version switch."""
     parsed_version: Optional[str]
-    """Semver parsed version."""
+    """Clean stringification of the version object for current schema."""
     is_compatible: str
     is_broken: bool
     last_modified: Optional[datetime.datetime]
@@ -25,6 +30,9 @@ class ToolCheckResult:
     def is_problem(self) -> bool:
         """Is this tool's state a problem?
         If it is needed for this OS, if it is incompatible or unavailable, it is a problem.
+
+        Returns:
+            bool: True if this tool's state is a problem, False otherwise.
         """
         missing_or_incompatible = self.is_compatible != "Compatible" or not self.is_available
         return self.is_needed_for_os and missing_or_incompatible
@@ -33,17 +41,22 @@ class ToolCheckResult:
 @dataclass
 class ToolAvailabilityResult:
     """
-    Dataclass for the result of checking tool availability.
+    Represents only if the tool is available or not.
     """
 
     is_available: bool
     is_broken: bool
     version: Optional[str]
+    """Desired version"""
     last_modified: Optional[datetime.datetime]
 
 
 @dataclass
 class CliToolConfig:
+    """
+    Represents what tool and what version the user wants to audit on their system.
+    """
+
     name: str
     version: Optional[str] = None
     version_snapshot: Optional[str] = None
@@ -54,26 +67,22 @@ class CliToolConfig:
     version_stamp: Optional[str] = None
     source: Optional[str] = None
 
+    def cache_hash(self) -> str:
+        """
+        Generate a hash for a CliToolConfig instance.
 
-def cache_hash(tool_config: CliToolConfig) -> str:
-    """
-    Generate a hash for a CliToolConfig instance.
+        Returns:
+            str: The hash of the tool configuration.
+        """
+        config_str = ""
+        for key, value in asdict(self).items():
+            config_str += f"{key}={value};"  # Concatenate key-value pairs
 
-    Args:
-        tool_config (CliToolConfig): The tool configuration.
-
-    Returns:
-        str: The hash of the tool configuration.
-    """
-    config_str = ""
-    for key, value in asdict(tool_config).items():
-        config_str += f"{key}={value};"  # Concatenate key-value pairs
-
-    # Use hashlib to compute an MD5 hash of the concatenated string
-    return hashlib.md5(config_str.encode()).hexdigest()  # nosec
+        # Use hashlib to compute an MD5 hash of the concatenated string
+        return hashlib.md5(config_str.encode()).hexdigest()  # nosec
 
 
 if __name__ == "__main__":
     # Example usage
     config = CliToolConfig(name="example_tool", version="1.0.0")
-    print(cache_hash(config))
+    print(config.cache_hash())